moved and updated tests #1279
