+++ test/Altinn.Platform.Authentication.Tests/Controllers/SystemRegisterControllerTests.cs
@@ -6,9 +6,8 @@ using System.Net;
 using System.Net.Http;
 using System.Net.Http.Headers;
 using System.Net.Http.Json;
-using System.Security.Policy;
+using System.Text;
 using System.Text.Json;
-using System.Threading;
 using System.Threading.Tasks;
 using Altinn.Authentication.Tests.Mocks;
 using Altinn.Authorization.ProblemDetails;
@@ -29,11 +28,7 @@ using Altinn.Platform.Authentication.Tests.RepositoryDataAccess;
 using Altinn.Platform.Authentication.Tests.Utils;
 using AltinnCore.Authentication.JwtCookie;
 using App.IntegrationTests.Utils;
-using Azure;
-using ICSharpCode.SharpZipLib.Zip;
 using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.Azure.KeyVault.Models;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Options;
@@ -53,7 +48,7 @@ namespace Altinn.Platform.Authentication.Tests.Controllers
         private readonly Mock<TimeProvider> timeProviderMock = new Mock<TimeProvider>();
         private readonly Mock<IGuidService> guidService = new Mock<IGuidService>();
         private readonly Mock<IEventsQueueClient> _eventQueue = new Mock<IEventsQueueClient>();
-        
+
         private readonly JsonSerializerOptions options = new()
         {
             PropertyNamingPolicy = JsonNamingPolicy.CamelCase
@@ -964,14 +959,14 @@ namespace Altinn.Platform.Authentication.Tests.Controllers
         public async Task SystemRegister_Update_System()
         {
             // Post original System
-            string dataFileName = "Data/SystemRegister/Json/SystemRegister.json";
+            const string dataFileName = "Data/SystemRegister/Json/SystemRegister.json";
             HttpResponseMessage response = await CreateSystemRegister(dataFileName);
             Assert.Equal(HttpStatusCode.OK, response.StatusCode);
 
             // Prepare updated system request
             string systemId = "991825827_the_matrix";
             HttpClient client = CreateClient();
-            string[] prefixes = { "altinn", "digdir" };
+            string[] prefixes = ["altinn", "digdir"];
             string token = PrincipalUtil.GetOrgToken("digdir", "991825827", "altinn:authentication/systemregister.admin", prefixes);
             client.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue("Bearer", token);
 
@@ -979,7 +974,7 @@ namespace Altinn.Platform.Authentication.Tests.Controllers
             StreamContent content = new StreamContent(dataStream);
             content.Headers.ContentType = new MediaTypeHeaderValue("application/json");
 
-            // Run update request with two new client_id's: removing one existing and adding two new ones
+            // Run update request with two new client_id's - removing one existing and adding two new ones
             HttpRequestMessage request = new(HttpMethod.Put, $"/authentication/api/v1/systemregister/vendor/{systemId}/");
             request.Content = content;
             HttpResponseMessage updateResponse = await client.SendAsync(request, HttpCompletionOption.ResponseContentRead);
@@ -995,7 +990,7 @@ namespace Altinn.Platform.Authentication.Tests.Controllers
 
             // Assert updates were made
             AssertionUtil.AssertRegisteredSystem(expectedRegisteredSystem, actualUpdatedSystem);
-            
+
             // Verify you can create new system with old (deleted) clientIds
             string filename = "Data/SystemRegister/Json/SystemRegisterClientIdsExist.json";
             HttpResponseMessage responseNewSystem = await CreateSystemRegister(filename);
@@ -1164,6 +1159,220 @@ namespace Altinn.Platform.Authentication.Tests.Controllers
             AltinnValidationError error = problemDetails.Errors.Single(e => e.ErrorCode == ValidationErrors.SystemRegister_AccessPackage_NotValid.ErrorCode);
             Assert.Equal("/registersystemrequest/accesspackages", error.Paths.Single(p => p.Equals("/registersystemrequest/accesspackages")));
             Assert.Equal("One or all the accesspackage(s) is not found in altinn's access packages or is not delegable because they are not part of REGN/REVI/Forretningsfører roller", error.Detail);
+        }
+
+        [Fact]
+        public async Task SystemRegister_Update_System_UnchangedClientId_Test()
+        {
+            // Prepare
+            const string systemId = "991825827_the_matrix";
+            List<string> clientIdsInFirstSystem = [Guid.NewGuid().ToString(), Guid.NewGuid().ToString(), Guid.NewGuid().ToString()];
+            RegisterSystemRequest originalSystem = CreateSystemRegisterRequest(systemId, clientIdsInFirstSystem);
+            HttpResponseMessage response = await CreateSystemRegister(originalSystem);
+            Assert.Equal(HttpStatusCode.OK, response.StatusCode);
+
+            // Keep using the same clientId, but update something else
+            RegisterSystemRequest updatedSystem = CreateSystemRegisterRequest(systemId, clientIdsInFirstSystem, true);
+            var resp = await PutSystemRegisterAsync(updatedSystem, systemId);
+            Assert.Equal(HttpStatusCode.OK, resp.StatusCode);
+
+            HttpResponseMessage getSystemResponse = await GetSystemRegister(systemId);
+            Assert.Equal(HttpStatusCode.OK, getSystemResponse.StatusCode);
+
+            // Assert new system contains the same clientId
+            RegisteredSystemResponse actualUpdatedSystem = JsonSerializer.Deserialize<RegisteredSystemResponse>(await getSystemResponse.Content.ReadAsStringAsync(), _options);
+            Assert.True(actualUpdatedSystem.ClientId.Count == clientIdsInFirstSystem.Count);
+            Assert.Contains(clientIdsInFirstSystem[0], actualUpdatedSystem.ClientId[0]);
+            Assert.True(actualUpdatedSystem.IsVisible);
+        }
+
+        [Fact]
+        public async Task SystemRegister_DuplicateClientIds_Test()
+        {
+            // Prepare
+            const string systemId = "991825827_the_matrix";
+            List<string> clientIdsInFirstSystem = ["456", "8910"];
+            RegisterSystemRequest originalSystem = CreateSystemRegisterRequest(systemId, clientIdsInFirstSystem);
+            HttpResponseMessage response = await CreateSystemRegister(originalSystem);
+            Assert.Equal(HttpStatusCode.OK, response.StatusCode);
+
+            // Duplicate clientIds
+            List<string> newClientIds = ["123", "123"];
+            RegisterSystemRequest updatedSystem = CreateSystemRegisterRequest(systemId, newClientIds);
+            var resp = await PutSystemRegisterAsync(updatedSystem, systemId);
+            Assert.Equal(HttpStatusCode.BadRequest, resp.StatusCode);
+
+            string content = await resp.Content.ReadAsStringAsync();
+            Assert.Contains(ValidationErrors.SystemRegister_Duplicate_ClientIds.Detail, content);
+        }
+
+        [Fact]
+        public async Task SystemRegister_UnmatchedRequestBodyAndSystemId()
+        {
+            // Prepare
+            const string systemId = "991825827_the_matrix";
+            List<string> clientIdsInFirstSystem = [Guid.NewGuid().ToString(), Guid.NewGuid().ToString(), Guid.NewGuid().ToString()];
+            RegisterSystemRequest originalSystem = CreateSystemRegisterRequest(systemId, clientIdsInFirstSystem);
+            HttpResponseMessage response = await CreateSystemRegister(originalSystem);
+            Assert.Equal(HttpStatusCode.OK, response.StatusCode);
+
+            List<string> newClientIds = [Guid.NewGuid().ToString(), Guid.NewGuid().ToString()];
+            RegisterSystemRequest updatedSystem = CreateSystemRegisterRequest(systemId, newClientIds);
+
+            // Expecting bad request here
+            var resp = await PutSystemRegisterAsync(updatedSystem, "991825827_does_not_match_request_bodys_system_id");
+            Assert.Equal(HttpStatusCode.BadRequest, resp.StatusCode);
+
+            // Read the content as string
+            string content = await resp.Content.ReadAsStringAsync();
+            Assert.Equal("SystemId in request body doesn't match systemId in Url", content);
+        }
+
+        [Fact]
+        public async Task SystemRegister_SystemNotFound_Test()
+        {
+            // Prepare
+            const string systemId = "991825827_the_matrix";
+            List<string> clientIdsInFirstSystem = [Guid.NewGuid().ToString(), Guid.NewGuid().ToString(), Guid.NewGuid().ToString()];
+            RegisterSystemRequest originalSystem = CreateSystemRegisterRequest(systemId, clientIdsInFirstSystem);
+            HttpResponseMessage response = await CreateSystemRegister(originalSystem);
+            Assert.Equal(HttpStatusCode.OK, response.StatusCode);
+
+            // Try to put on system that does not exist
+            const string putSystemId = "SystemIdThatDoesNotExist";
+            List<string> newClientIds = [Guid.NewGuid().ToString(), Guid.NewGuid().ToString()];
+            RegisterSystemRequest updatedSystem = CreateSystemRegisterRequest(putSystemId, newClientIds);
+
+            // Expecting bad request here
+            var resp = await PutSystemRegisterAsync(updatedSystem, putSystemId);
+            Assert.Equal(HttpStatusCode.NotFound, resp.StatusCode);
+
+            // Read the content as string
+            string content = await resp.Content.ReadAsStringAsync();
+            Assert.Equal($"System with ID '{putSystemId}' not found.", content);
+        }
+
+        [Fact]
+        public async Task SystemRegister_ClientIdExists_Test()
+        {
+            // Prepare
+            const string systemId = "991825827_the_matrix";
+            const string systemIdSecondSystem = "991825827_snowman";
+
+            List<string> clientIdsInFirstSystem = [Guid.NewGuid().ToString(), Guid.NewGuid().ToString(), Guid.NewGuid().ToString()];
+            HttpResponseMessage responseFirst = await CreateAndAssertSystemAsync(systemId, clientIdsInFirstSystem);
+            Assert.Equal(HttpStatusCode.OK, responseFirst.StatusCode);
+
+            List<string> clientIdsSecondSystem = ["ClientIdForSecondSystem"];
+            await CreateAndAssertSystemAsync(systemIdSecondSystem, clientIdsSecondSystem);
+
+            // Running update with one new clientId and also one old from a second system
+            List<string> newClientIds = ["NewClientIdToUpdate", clientIdsSecondSystem[0]];
+            RegisterSystemRequest updatedSystem = CreateSystemRegisterRequest(systemId, newClientIds);
+
+            // Expecting bad request here
+            HttpResponseMessage resp = await PutSystemRegisterAsync(updatedSystem, systemId);
+            Assert.Equal(HttpStatusCode.BadRequest, resp.StatusCode);
+            string content = await resp.Content.ReadAsStringAsync();
+            Assert.Contains(ValidationErrors.SystemRegister_ClientID_Exists.Detail, content);
+        }
+
+        private async Task<HttpResponseMessage> CreateAndAssertSystemAsync(string systemId, List<string> clientIds)
+        {
+            var request = CreateSystemRegisterRequest(systemId, clientIds);
+            HttpResponseMessage response = await CreateSystemRegister(request);
+            return response;
+        }
+
+        private async Task<HttpResponseMessage> CreateSystemRegister(RegisterSystemRequest registerSystemRequest)
+        {
+            HttpClient client = CreateClient();
+
+            string[] prefixes = ["altinn", "digdir"];
+            string token = PrincipalUtil.GetOrgToken(
+                "digdir",
+                "991825827",
+                "altinn:authentication/systemregister.admin",
+                prefixes);
+
+            client.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue("Bearer", token);
+
+            string json = JsonSerializer.Serialize(registerSystemRequest, _options);
+            HttpContent content = new StringContent(json, Encoding.UTF8, "application/json");
+
+            HttpRequestMessage request = new(HttpMethod.Post, "/authentication/api/v1/systemregister/vendor/")
+            {
+                Content = content
+            };
+
+            HttpResponseMessage response = await client.SendAsync(request, HttpCompletionOption.ResponseHeadersRead);
+            return response;
+        }
+
+        private async Task<HttpResponseMessage> PutSystemRegisterAsync(RegisterSystemRequest updateDto, string systemIdPath)
+        {
+            HttpClient client = CreateClient();
+
+            string[] prefixes = ["altinn", "digdir"];
+            string token = PrincipalUtil.GetOrgToken("digdir", "991825827", "altinn:authentication/systemregister.admin", prefixes);
+            client.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue("Bearer", token);
+
+            var json = JsonSerializer.Serialize(updateDto, options);
+            using var content = new StringContent(json, Encoding.UTF8, "application/json");
+
+            var request = new HttpRequestMessage(HttpMethod.Put, $"/authentication/api/v1/systemregister/vendor/{systemIdPath}/")
+            {
+                Content = content
+            };
+
+            return await client.SendAsync(request, HttpCompletionOption.ResponseHeadersRead);
+        }
+
+        private static RegisterSystemRequest CreateSystemRegisterRequest(string systemId, List<string> clientIds, bool isVisible = false)
+        {
+            return new RegisterSystemRequest
+            {
+                Id = systemId,
+                Vendor = new VendorInfo
+                {
+                    Authority = "iso6523-actorid-upis",
+                    ID = "0192:991825827"
+                },
+                Name = new Dictionary<string, string>
+                {
+                    { "nb", "The Matrix" },
+                    { "en", "The Matrix" },
+                    { "nn", "The Matrix" }
+                },
+                Description = new Dictionary<string, string>
+                {
+                    { "nb", "Test system for Put" },
+                    { "en", "Test system for Put" },
+                    { "nn", "Test system for Put" }
+                },
+                Rights = new List<Right>
+                {
+                    new()
+                    {
+                        Resource = new List<AttributePair>
+                        {
+                            new()
+                            {
+                                Id = "urn:altinn:resource",
+                                Value = "ske-krav-og-betalinger"
+                            }
+                        }
+                    }
+                },
+                ClientId = clientIds,
+                AllowedRedirectUrls = new List<Uri>
+                {
+                    new("https://vg.no"),
+                    new("https://nrk.no"),
+                    new("https://altinn.no")
+                },
+                IsVisible = isVisible
+            };
         }
 
         private void SetupDateTimeMock()
