+++ test/Altinn.Platform.Authentication.Tests/Data/SystemRegister/Json/SystemRegisterUpdateRequest.json
@@ -31,7 +31,6 @@
   ],
   "softDeleted": false,
   "clientId": [
-    "32ef65ac-6e62-498d-880f-76c85c2052ae",
     "UPDATED CLIENT ID",
     "ANOTHER UPDATED CLIENT ID",
     "GUID_SOMETHING"
